Add codecov and prepare for css parsing
