+++ .circleci/config.yml
@@ -1,8 +1,12 @@
-version: 2
+version: 2.1
+orbs:
+  codecov: codecov/codecov@1.0.2
 jobs:
   build:
     docker:
     - image: cirrusci/flutter
     steps:
     - checkout
-    - run: flutter test
+    - run: flutter test --coverage
+    - codecov/upload:
+        file: coverage/lcov.info
