+++ lib/content_element.dart
@@ -65,8 +65,8 @@ class ImageContentElement extends ContentElement {
     } else {
       return Image.network(src);
     }
-    //TODO alt text
-    //TODO precacheImage
+    //TODO(Sub6Resources): alt text
+    //TODO(Sub6Resources): precacheImage
   }
 }
 
@@ -90,7 +90,7 @@ class AudioContentElement extends ContentElement {
 
   @override
   Widget toWidget() {
-    //TODO
+    //TODO(Sub6Resources)
     return Container(padding: const EdgeInsets.all(24), child: Text("AUDIO"));
   }
 }
@@ -117,7 +117,7 @@ class VideoContentElement extends ContentElement {
 
   @override
   Widget toWidget() {
-    //TODO
+    //TODO(Sub6Resources)
     return Container(padding: const EdgeInsets.all(24), child: Text("AUDIO"));
   }
 }
